requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
python-dotenv==0.21.1
SQLAlchemy==1.4.46
SQLAlchemy_Utils==0.39.0
